lib: Rename aio.abc to aio.base
Fixes error with ruff 0.6.7:

> lib/aio/abc.py:1:1: A005 Module `abc` is shadowing a Python builtin module
